添加：大部分图片已添加 测试：登录请求发送成功

+++ SchoolMarket/ViewController.m
@@ -70,7 +70,7 @@
 }
 //点击事件 （重用bug）
 - (void)click:(UIButton *)button {
-    NSLog(@"%d",button.tag);
+    NSLog(@"%ld",(long)button.tag);
     
     //获取button所在的cell
     CommCell *cell = (CommCell *)[button superview];
